fix: refine desktop toolbar and sidebar controls
